修改第一版

+++ app/views.py
@@ -5,7 +5,7 @@ from json.decoder import JSONObject
 from urllib.parse import parse_qs
 
 from django.core.cache import cache
-from django.http import JsonResponse
+from django.http import JsonResponse, HttpResponse
 from django.shortcuts import render, redirect
 
 # Create your views here.
@@ -36,7 +36,6 @@ def index(request):
 
         response_data['user'] = user
 
-
     return render(request,'index.html',context=response_data)
 
 @csrf_exempt
@@ -46,6 +45,7 @@ def login(request):
     elif request.method == 'POST':
         phone = request.POST.get('phone')
         password = request.POST.get('password')
+        back = request.COOKIES.get('back')
 
         users = User.objects.filter(phone=phone)
         if users.exists():  # 存在
@@ -53,12 +53,12 @@ def login(request):
             if user.password == generate_password(password):  # 验证通过
                 # 更新token
                 token = generate_token()
-
                 # 状态保持
                 cache.set(token, user.id, 60 * 60 * 24 * 7)
-
                 # 传递客户端
                 request.session['token'] = token
+                if back == 'goods':
+                    return redirect('fashion:goods')
                 return render(request,'index.html')
             else:  # 密码错误
                 return render(request, 'login.html')
@@ -68,7 +68,6 @@ def login(request):
 
 def logout(request):
     request.session.flush()
-
     return redirect('fashion:index')
 
 
@@ -77,20 +76,16 @@ def generate_password(param):
     md5.update(param.encode('utf-8'))
     return md5.hexdigest()
 
-
 def generate_token():
     temp = str(time.time()) + str(random.random())
     md5 = hashlib.md5()
     md5.update(temp.encode('utf-8'))
     return md5.hexdigest()
 
-
 def register(request):
     if request.method == 'GET':
         return render(request, 'register.html')
     elif request.method == 'POST':
-
-
         # 获取数据
         phone = request.POST.get('phone')
         passoword = generate_password(request.POST.get('password'))
@@ -101,7 +96,6 @@ def register(request):
         user.password = passoword
         user.name=name
         user.save()
-
         # 状态保持
         token = generate_token()
         # key-value  >>  token:userid
@@ -111,24 +105,24 @@ def register(request):
         print(token)
         return redirect('fashion:index')
 
-
 def goods(request):
     index = int(request.COOKIES.get('index'))
     goods=Goods.objects.all()
     good=goods[index]
     token = request.session.get('token')
     userid = cache.get(token)
-    user = User.objects.get(pk=userid)
-    cars = Car.objects.filter(user=user)
+    response_data={
+            'good': good,
+            'user': None,
+            'cars': None,
+    }
+    if userid:
+        user = User.objects.get(pk=userid)
+        cars = Car.objects.filter(user=user)
+        response_data['user']=user
+        response_data['cars']=cars
 
-    return render(request,'goods.html',context={
-        'good':good,
-        'user':user,
-        'cars':cars,
-    })
-
-
-
+    return render(request,'goods.html',context=response_data)
 
 def list(request):
     goods=Goods.objects.all()
@@ -136,22 +130,25 @@ def list(request):
     token = request.session.get('token')
     userid = cache.get(token)
     # index = int(request.COOKIES.get('index'))
-    user = User.objects.get(pk=userid)
+    response_data = {
+        'goods': goods,
+        'user': None,
+    }
+    if userid:
+        user = User.objects.get(pk=userid)
+        print(user.name)
+        response_data['user'] = user
     # 根据index 获取 对应的 分类ID
     # categoryid = foodtypes[index].typeid
     # 根据 分类ID 获取对应分类信息
-    # goods_list = Goods.objects.filter(categoryid=categoryid)
-    # a=index+1000
     # go=goods[index]
     # good=Goods.objects.get(id=go)
     # print(index)
     # print(go,'就是这个商品')
-    return render(request,'list.html',context={
-        'goods':goods,
-        'user':user,
-    })
+    return render(request,'list.html',context=response_data)
 
 def car(request):
+
     token = request.session.get('token')
     userid = cache.get(token)
     if userid:  # 有登录才显示
@@ -162,7 +159,6 @@ def car(request):
         for car in cars:
             if not car.isselect:
                 isall = False
-
         return render(request, 'car.html', context={
             'cars': cars,
             'isall':isall,
@@ -170,7 +166,6 @@ def car(request):
         })
     else:   # 未登录不显示
         return render(request, 'login.html')
-
 
 def addcar(request):
     # 获取token
@@ -201,13 +196,9 @@ def addcar(request):
 
         response_data['status'] = 1
         response_data['number'] = car.number
-        response_data['msg'] = '添加 {} 购物车成功: {}'.format(car.goods.wen, car.number)
+        response_data['msg'] = '成功添加 {} 到购物车: {}'.format(car.goods.wen, car.number)
 
         return JsonResponse(response_data)
-
-    # 未登录
-    # 因为是ajax操作，所以重定向是不可以的!
-    # return redirect('axf:login')
 
     response_data['status'] = -1
     response_data['msg'] = '请登录后操作'
@@ -245,7 +236,6 @@ def changecarselect(request):
     car = Car.objects.get(pk=carid)
     car.isselect = not car.isselect
     car.save()
-
     response_data = {
         'msg': '状态修改成功',
         'status': 1,
@@ -281,7 +271,7 @@ def changecarall(request):
 
 
 def generate_identifier():
-    temp = str(time.time()) + str(random.randrange(1000, 10000))
+    temp = str(time.time()) + str(random.randrange(100, 10000))
     return temp
 
 
@@ -299,36 +289,38 @@ def generateorder(request):
 
     # 订单商品(购物车中选中)
     cars = user.car_set.filter(isselect=True)
+    if cars:
+        for car in cars:
+            orderGoods = OrderGoods()
+            orderGoods.order = order
+            orderGoods.goods = car.goods
+            orderGoods.number = car.number
+            orderGoods.save()
+            # 购物车中移除
+            car.delete()
 
-    for car in cars:
-        orderGoods = OrderGoods()
-        orderGoods.order = order
-        orderGoods.goods = car.goods
-        orderGoods.number = car.number
-        orderGoods.save()
-
-        # 购物车中移除
-        car.delete()
-
-    # response_data = {
-    #     'msg': '生成订单',
-    #     'status': 1,
-    #     'identifier': order.identifier
-    print(order.ordergoods_set.all().first())
-    return render(request, 'orderdetail.html', context={'order': order})
-
+        # response_data = {
+        #     'msg': '生成订单',
+        #     'status': 1,
+        #     'identifier': order.identifier
+        print(order.ordergoods_set.all().first())
+        return render(request, 'orderdetail.html', context={
+            'order': order,
+            'user':user
+        })
+    else:
+        return HttpResponse("还没有选择商品")
 
 def orderlist(request):
     token = request.session.get('token')
     userid = cache.get(token)
     user = User.objects.get(pk=userid)
-
     orders = user.order_set.all()
-
     # status_list = ['未付款', '待发货', '待收货', '待评价', '已评价']
-
-    return render(request, 'orderlist.html', context={'orders': orders})
-
+    return render(request, 'orderlist.html', context={
+        'orders': orders,
+        'user':user,
+    })
 
 def orderdetail(request, identifier):
     order = Order.objects.filter(identifier=identifier).first()
@@ -338,32 +330,26 @@ def orderdetail(request, identifier):
 
 def returnurl(request):
     return redirect('fashion:index')
+
 @csrf_exempt
 def appnotifyurl(request):
     if request.method == 'POST':
         # 获取到参数
         body_str = request.body.decode('utf-8')
-
         # 通过parse_qs函数
         post_data = parse_qs(body_str)
-
         # 转换为字典
         post_dic = {}
         for k,v in post_data.items():
             post_dic[k] = v[0]
-
         # 获取订单号
         out_trade_no = post_dic['out_trade_no']
-
         # 更新状态
         Order.objects.filter(identifier=out_trade_no).update(status=1)
-
-
-    return JsonResponse({'msg':'success'})
+    return JsonResponse({'msg':'支付成功'})
 
 
 def pay(request):
-    # print(request.GET.get('orderid'))
 
     orderid = request.GET.get('orderid')
     order = Order.objects.get(pk=orderid)
@@ -374,7 +360,7 @@ def pay(request):
 
     # 支付地址信息
     data = alipay.direct_pay(
-        subject='就是要买卖买......', # 显示标题
+        subject='就是要买买买......', # 显示标题
         out_trade_no=order.identifier,    # 商品 订单号
         total_amount=str(sum),   # 支付金额
         return_url='http://120.79.59.213/fashion/returnurl/'
